refactored code for greater organization before implementing redux
