+++ app/App.js
@@ -1,0 +1,16 @@
+'use strict';
+
+import React, { Component } from 'react';
+import TabBar from './components/TabBar.js';
+
+class App extends Component {
+  constructor(props) {
+    super(props)
+  }  
+
+  render() {
+    return <TabBar/>
+  }
+}
+
+module.exports = App;
