+++ app/components/chart.js
@@ -13,7 +13,7 @@ import {
   ActivityIndicator
 } from 'react-native';
 import styles from './styles';
-import { getDbData, formatElapTime } from '../utils';
+import { getDbData, formatElapTime } from '../../utils';
 
 var ds = new ListView.DataSource({rowHasChanged: (r1, r2) => r1 !== r2});
 var greatestPercentage = 0;
