+++ app/components/log/logDetailView.js
@@ -10,7 +10,7 @@ import {
   getDbData,
   formatToTime,
   formatToDate
-} from '../../utils'
+} from '../../../utils'
 
 var ds = new ListView.DataSource({rowHasChanged: (r1, r2) => r1 !== r2});
