+++ index.ios.js
@@ -7,25 +7,8 @@
 //  */
 //
 
-import React, { Component } from 'react';
-import {
-  AsyncStorage,
-  AppRegistry,
-  View
-} from 'react-native';
-import TabBar from './components/tabBar';
-import styles from './components/styles';
-import {SERVER_ROUTE} from './server/env/development';
+import { AppRegistry } from 'react-native';
+import App from './app/App';
 import db from './database';
 
-class capstone extends Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return <TabBar/>
-  }
-}
-
-AppRegistry.registerComponent('capstone', () => capstone);
+AppRegistry.registerComponent('capstone', () => App);
